# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Library Management System API",
    "description": "A **Mini Library Management System** built with FastAPI, PostgreSQL, and OpenAI.\n\n## Authentication\n\nAll protected endpoints require a **Bearer JWT** obtained through the OAuth login flow:\n\n1. Open `GET /api/v1/auth/login/google` (or `/github`) in your **browser**.\n2. Complete the OAuth consent screen.\n3. You are redirected to the frontend with `?token=<jwt>` in the URL.\n4. Copy the token and paste it into the **Authorize** dialog (🔒 button above).\n\nEvery protected endpoint then works directly from this Swagger UI.\n\n## Roles & Permissions\n\n| Role | Capabilities |\n|------|--------------|\n| **Admin** | Full access including user management |\n| **Librarian** | Create / edit / delete books; manage all loans |\n| **Member** | Browse and search books; borrow and return **own** loans |\n\nNew OAuth users start as **Member**. An Admin must promote them via `PATCH /api/v1/admin/users/{id}/role`.\n\n## AI Features\n\nAll three AI endpoints degrade gracefully — they never return an error due to missing AI configuration. Check the `source` field (`\"openai\"` vs `\"fallback\"`) to see which path was taken.\n",
    "contact": {
      "name": "Library Management System",
      "url": "https://github.com/maherabbass/Library-Management-System"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/books/enrich": {
      "post": {
        "tags": [
          "books",
          "ai"
        ],
        "summary": "Generate AI metadata (no DB write)",
        "description": "Generates a **summary**, **tags**, and **keywords** for a book using OpenAI.\n\nThis endpoint is **preview-only** — nothing is written to the database. The librarian reviews the output and then includes it when calling `POST /books` (create) or `PUT /books/{id}` (update).\n\n**Fallback:** if `OPENAI_API_KEY` is not set or the AI call fails, a deterministic heuristic is used instead. Check `source` in the response.\n\n**Requires:** Librarian or Admin role.",
        "operationId": "enrich_book_endpoint_api_v1_books_enrich_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnrichRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Generated metadata — summary, tags, keywords, and the source that produced them.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrichResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired Bearer token."
          },
          "403": {
            "description": "Forbidden — Librarian or Admin role required."
          },
          "422": {
            "description": "Validation error — `title` and `author` are required."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/books/ai-search": {
      "get": {
        "tags": [
          "books",
          "ai"
        ],
        "summary": "Semantic search with embeddings",
        "description": "Ranks books by **semantic similarity** to a natural-language query using OpenAI text embeddings and cosine similarity.\n\nThis endpoint is **public** — no authentication required.\n\n**Fallback:** when `OPENAI_API_KEY` is not configured, a standard ILIKE keyword search is performed instead. The `source` field indicates which method was used.\n\n**Example queries:**\n- `books about dystopian futures`\n- `science fiction with strong female protagonists`\n- `classic literature set in 19th century England`",
        "operationId": "ai_search_endpoint_api_v1_books_ai_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Natural-language search query.",
              "examples": [
                "books about space exploration"
              ],
              "title": "Q"
            },
            "description": "Natural-language search query."
          },
          {
            "name": "top_k",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "Maximum number of results to return (1–50).",
              "default": 10,
              "title": "Top K"
            },
            "description": "Maximum number of results to return (1–50)."
          }
        ],
        "responses": {
          "200": {
            "description": "Ranked list of matching books with search metadata.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AISearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/books/ask": {
      "post": {
        "tags": [
          "books",
          "ai"
        ],
        "summary": "Ask the Library assistant",
        "description": "Ask a natural-language question and receive an answer grounded **exclusively** in books that exist in the database.\n\nThe assistant never invents titles, authors, or availability — every claim is backed by a real database record returned in the `books` array.\n\n**Fallback:** when `OPENAI_API_KEY` is not configured, a pre-built template answer is returned using the matched books.\n\n**Example questions:**\n- `Do you have any books by Isaac Asimov?`\n- `What science fiction books are available?`\n- `Can you recommend something about machine learning?`\n\n**Requires:** any authenticated user (Member, Librarian, or Admin).",
        "operationId": "ask_library_endpoint_api_v1_books_ask_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Grounded answer with the database records that support it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AskResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired Bearer token."
          },
          "422": {
            "description": "Validation error — `question` is required and must be ≤500 characters."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/books": {
      "get": {
        "tags": [
          "books"
        ],
        "summary": "List books",
        "description": "Returns a **paginated, filterable** list of books from the catalogue.\n\nThis endpoint is **public** — no authentication required.\n\n**Filters (all optional, combinable):**\n\n| Parameter | Behaviour |\n|-----------|----------|\n| `q` | Case-insensitive substring match across title, author, ISBN, and description |\n| `author` | Case-insensitive substring match on the author field only |\n| `tag` | Exact match against the tags array |\n| `status` | `AVAILABLE` or `BORROWED` |\n\nResults are ordered by `created_at` descending (newest first).",
        "operationId": "list_books_endpoint_api_v1_books_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Free-text search across title, author, ISBN, and description (case-insensitive).",
              "examples": [
                "Frank Herbert"
              ],
              "title": "Q"
            },
            "description": "Free-text search across title, author, ISBN, and description (case-insensitive)."
          },
          {
            "name": "author",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by author name (case-insensitive substring match).",
              "examples": [
                "Tolkien"
              ],
              "title": "Author"
            },
            "description": "Filter by author name (case-insensitive substring match)."
          },
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by exact tag value.",
              "examples": [
                "science-fiction"
              ],
              "title": "Tag"
            },
            "description": "Filter by exact tag value."
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/BookStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by availability status: `AVAILABLE` or `BORROWED`.",
              "title": "Status"
            },
            "description": "Filter by availability status: `AVAILABLE` or `BORROWED`."
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number (1-based).",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number (1-based)."
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Number of items per page (1–100).",
              "default": 20,
              "title": "Page Size"
            },
            "description": "Number of items per page (1–100)."
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated book list with total count and page metadata.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "books"
        ],
        "summary": "Create a book",
        "description": "Adds a new book to the catalogue.\n\nThe book is created with status `AVAILABLE`. Only `title` and `author` are required — all other fields are optional.\n\n**Tip:** call `POST /books/enrich` first to generate AI-suggested tags and description, then include them in this request.\n\n**Requires:** Librarian or Admin role.",
        "operationId": "create_book_endpoint_api_v1_books_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The newly created book record.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired Bearer token."
          },
          "403": {
            "description": "Forbidden — Librarian or Admin role required."
          },
          "409": {
            "description": "A book with the same ISBN already exists."
          },
          "422": {
            "description": "Validation error — check `title` and `author` are non-empty."
          }
        }
      }
    },
    "/api/v1/books/{book_id}": {
      "get": {
        "tags": [
          "books"
        ],
        "summary": "Get a book",
        "description": "Returns the full details of a single book by its UUID.\n\nThis endpoint is **public** — no authentication required.",
        "operationId": "get_book_endpoint_api_v1_books__book_id__get",
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Book Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The requested book record.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookResponse"
                }
              }
            }
          },
          "404": {
            "description": "Book not found."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "books"
        ],
        "summary": "Update a book",
        "description": "Partially updates a book's metadata using the provided fields.\n\nOnly fields included in the request body are changed — omitted fields keep their current values (partial update / PATCH semantics on a PUT).\n\n**Requires:** Librarian or Admin role.",
        "operationId": "update_book_endpoint_api_v1_books__book_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Book Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated book record.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired Bearer token."
          },
          "403": {
            "description": "Forbidden — Librarian or Admin role required."
          },
          "404": {
            "description": "Book not found."
          },
          "422": {
            "description": "Validation error — check field types and constraints."
          }
        }
      },
      "delete": {
        "tags": [
          "books"
        ],
        "summary": "Delete a book",
        "description": "Permanently removes a book from the catalogue.\n\n> **Warning:** this operation is irreversible. Associated loan history is also deleted via cascade.\n\n**Requires:** Librarian or Admin role.",
        "operationId": "delete_book_endpoint_api_v1_books__book_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Book Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No content — the book was successfully deleted."
          },
          "401": {
            "description": "Missing, invalid, or expired Bearer token."
          },
          "403": {
            "description": "Forbidden — Librarian or Admin role required."
          },
          "404": {
            "description": "Book not found."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AISearchResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BookResponse"
            },
            "type": "array",
            "title": "Items",
            "description": "Books ranked by semantic similarity to the query (most similar first)."
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Number of books returned."
          },
          "source": {
            "type": "string",
            "enum": [
              "openai",
              "fallback"
            ],
            "title": "Source",
            "description": "`openai` — results ranked by cosine similarity of OpenAI embeddings. `fallback` — `OPENAI_API_KEY` is not set; ILIKE keyword search was used instead."
          },
          "query": {
            "type": "string",
            "title": "Query",
            "description": "The original query string that was searched."
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "source",
          "query"
        ],
        "title": "AISearchResponse",
        "example": {
          "items": [],
          "query": "books about space exploration and survival",
          "source": "openai",
          "total": 3
        }
      },
      "AskRequest": {
        "properties": {
          "question": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Question",
            "description": "A natural-language question about the library collection.",
            "examples": [
              "Do you have any books about artificial intelligence?"
            ]
          }
        },
        "type": "object",
        "required": [
          "question"
        ],
        "title": "AskRequest",
        "example": {
          "question": "Do you have any books about artificial intelligence?"
        }
      },
...
      "BookStatus": {
        "type": "string",
        "enum": [
          "AVAILABLE",
          "BORROWED"
        ],
        "title": "BookStatus"
      },
...
      "EnrichRequest": {
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "title": "Title",
            "description": "Book title to enrich.",
            "examples": [
              "Dune"
            ]
          },
          "author": {
            "type": "string",
            "minLength": 1,
            "title": "Author",
            "description": "Primary author of the book.",
            "examples": [
              "Frank Herbert"
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Optional synopsis or notes that help the AI generate richer metadata.",
            "examples": [
              "A science fiction epic set on the desert planet Arrakis."
            ]
          }
        },
        "type": "object",
        "required": [
          "title",
          "author"
        ],
        "title": "EnrichRequest",
        "example": {
          "author": "Frank Herbert",
          "description": "A science fiction epic set on the desert planet Arrakis.",
          "title": "Dune"
        }
      },
      "EnrichResponse": {
        "properties": {
          "summary": {
            "type": "string",
            "title": "Summary",
            "description": "AI-generated one-paragraph summary of the book. When `source` is `fallback` this equals the provided description."
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags",
            "description": "Suggested classification tags (genre, topic, era, …)."
          },
          "keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Keywords",
            "description": "Key terms extracted from the title/author/description for search indexing."
          },
          "source": {
            "type": "string",
            "enum": [
              "openai",
              "fallback"
            ],
            "title": "Source",
            "description": "`openai` — result was generated by the configured OpenAI model. `fallback` — `OPENAI_API_KEY` is not set or an error occurred; deterministic heuristics were used instead."
          }
        },
        "type": "object",
        "required": [
          "summary",
          "tags",
          "keywords",
          "source"
        ],
        "title": "EnrichResponse",
        "example": {
          "keywords": [
            "dune",
            "arrakis",
            "spice",
            "paul",
            "atreides",
            "herbert"
          ],
          "source": "openai",
          "summary": "A science fiction epic following Paul Atreides on the desert planet Arrakis, where control of the spice melange determines the fate of the universe.",
          "tags": [
            "science-fiction",
            "epic",
            "classic",
            "space-opera"
          ]
        }
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token obtained from `GET /api/v1/auth/callback/{provider}`. Obtain it by completing the OAuth login flow and copying the `token` query parameter from the redirect URL."
      }
    }
  },
...
}